+++ metalearners/tlearner.py
@@ -151,23 +151,23 @@ class TLearner(_ConditionalAverageOutcomeMetaLearner):
         )
 
     @classmethod
-    def necessary_onnx_models(cls) -> set[str]:
+    def _necessary_onnx_models(cls) -> set[str]:
         return {VARIANT_OUTCOME_MODEL}
 
-    @copydoc(MetaLearner.build_onnx, sep="")
-    def build_onnx(self, models: Mapping[str, Sequence], output_name: str = "tau"):
+    @copydoc(MetaLearner._build_onnx, sep="")
+    def _build_onnx(self, models: Mapping[str, Sequence], output_name: str = "tau"):
         """In the TLearner case, the necessary models are:
 
         * ``"variant_outcome_model"``
         """
-        warning_experimental_feature("build_onnx")
+        warning_experimental_feature("_build_onnx")
         check_spox_installed()
         import spox.opset.ai.onnx.v21 as op
         from onnx.checker import check_model
         from spox import build, inline
 
         self._validate_feature_set_none()
-        self._validate_onnx_models(models, self.necessary_onnx_models())
+        self._validate_onnx_models(models, self._necessary_onnx_models())
 
         input_dict = infer_input_dict(models[VARIANT_OUTCOME_MODEL][0])