+++ tests/test_tlearner.py
@@ -116,7 +116,7 @@ def test_tlearner_onnx(
         )
         onnx_models.append(onnx_model)
 
-    final = ml.build_onnx({VARIANT_OUTCOME_MODEL: onnx_models})
+    final = ml._build_onnx({VARIANT_OUTCOME_MODEL: onnx_models})
     sess = rt.InferenceSession(
         final.SerializeToString(), providers=rt.get_available_providers()
     )