+++ CHANGELOG.rst
@@ -12,11 +12,11 @@ Changelog
 
 **New features**
 
-* Add ``build_onnx`` to :class:`metalearners.MetaLearner` abstract class and implement it
+* Add ``_build_onnx`` to :class:`metalearners.MetaLearner` abstract class and implement it
   for :class:`metalearners.TLearner`, :class:`metalearners.XLearner`, :class:`metalearners.RLearner`
   and :class:`metalearners.DRLearner`.
 
-* Add ``necessary_onnx_models`` to :class:`metalearners.MetaLearner`.
+* Add ``_necessary_onnx_models`` to :class:`metalearners.MetaLearner`.
 
 0.8.0 (2024-07-22)
 ------------------